Patch steal_config_maybe

+++ neuromation/cli/utils.py
@@ -735,8 +735,10 @@ def pager_maybe(
 def steal_config_maybe(dst_path: pathlib.Path) -> None:
     if NEURO_STEAL_CONFIG in os.environ:
         src = pathlib.Path(os.environ[NEURO_STEAL_CONFIG])
+        if not src.exists():
+            return
         dst = Factory(dst_path).path
-        dst.mkdir(mode=0o700)
+        dst.mkdir(mode=0o700, exist_ok=True)
         for f in src.iterdir():
             target = dst / f.name
             shutil.copy(f, target)
